application/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
appdirs==1.4.4
asgiref==3.6.0
astor==0.8.1
async-generator==1.10
async-timeout==4.0.3
attrs==22.1.0
beautifulsoup4==4.11.1
black==22.12.0
bleach==6.0.0
blessed==1.20.0
boto3==1.28.62
botocore==1.31.62
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.3.0
cement==2.8.2
certifi==2022.9.24
cffi @ file:///croot/cffi_1670423208954/work
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
colorama==0.4.3
conda==23.1.0
conda-content-trust @ file:///tmp/abs_5952f1c8-355c-4855-ad2e-538535021ba5h26t22e5/croots/recipe/conda-content-trust_1658126371814/work
conda-package-handling @ file:///croot/conda-package-handling_1666940373510/work
cryptography @ file:///croot/cryptography_1665612644927/work
distlib==0.3.6
Django==4.2
docopt==0.6.2
docutils==0.19
exceptiongroup==1.0.0
feedparser==6.0.10
filelock==3.8.0
Flask==2.2.3
fpdf==1.7.2
frozenlist==1.4.0
fsspec==2023.9.0
future==0.16.0
git-python==1.0.3
gitdb==4.0.10
GitPython==3.1.36
google-auth==2.16.2
google-auth-oauthlib==1.0.0
gspread==5.11.3
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.16.4
idna @ file:///croot/idna_1666125576474/work
importlib-metadata==6.8.0
iniconfig==2.0.0
inquirer==3.1.3
itsdangerous==2.1.2
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==3.2.0
keyring==23.13.1
Kivy==2.2.1
Kivy-Garden==0.1.5
kivymd==1.1.1
litellm==0.1.610
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
more-itertools==9.1.0
multidict==6.0.4
mypy-extensions==0.4.3
mysql-connector-python==9.0.0
numpy @ file:///croot/numpy_and_numpy_base_1672336185480/work
oauth2client==4.1.3
oauthlib==3.2.2
open-interpreter==0.1.3
openai==0.27.10
outcome==1.2.0
packaging==23.1
pathspec==0.9.0
pdfkit==1.0.0
pika==1.3.2
Pillow==10.0.0
pkginfo==1.9.6
platformdirs==2.5.3
pluggy==1.5.0
prettytable==3.7.0
prometheus-client==0.16.0
psutil==5.9.5
psycopg2-binary==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycosat @ file:///croot/pycosat_1666805502580/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.15.1
PyJWT==2.6.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.1.1
pyrsistent==0.19.2
PySocks==1.7.1
pytest==8.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
pytz==2022.7.1
PyYAML==6.0.1
readchar==4.0.5
readme-renderer==37.3
regex==2023.8.8
reportlab==4.2.0
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.5.2
rsa==4.9
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
s3transfer==0.7.0
SecretStorage==3.3.3
selenium==4.5.0
semantic-version==2.8.5
sgmllib3k==1.0.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
system-data==1.0
tabulate==0.9.0
termcolor==1.1.0
texttable==1.6.4
tiktoken==0.4.0
tokentrim==0.1.10
tomli==2.0.1
toolz @ file:///croot/toolz_1667464077321/work
tqdm @ file:///opt/conda/conda-bld/tqdm_1664392687731/work
trio==0.22.0
trio-websocket==0.9.2
twilio==7.16.5
twine==4.0.2
typing_extensions==4.7.1
urllib3 @ file:///croot/urllib3_1670526988650/work
virtualenv==20.21.0
wcwidth==0.1.9
webbot==0.34
webencodings==0.5.1
websocket-client==0.59.0
Werkzeug==2.2.3
wget==3.2
wsproto==1.2.0
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.15.0
